pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jmd-mcp-sql"
version = "0.10.0"
description = "JMD-based MCP server for SQLite — natural language database interface"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Andreas Ostermeyer", email = "andreas@ostermeyer.de" }]
requires-python = ">=3.10"
keywords = ["jmd", "mcp", "sqlite", "llm", "natural-language", "database"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "mcp>=1.0",
    "jmd-format>=0.4",
]

[project.urls]
Homepage = "https://github.com/ostermeyer/jmd-spec"
Repository = "https://github.com/ostermeyer/jmd-mcp-sql"

[project.scripts]
jmd-mcp-sql = "jmd_mcp_sql.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["jmd_mcp_sql*"]

[tool.setuptools.package-data]
jmd_mcp_sql = ["*.sql"]

[tool.uv.sources]
jmd-format = { path = "../jmd-impl", editable = true }

[tool.ruff]
line-length = 80
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "D",   # pydocstyle (Google convention)
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
strict = true
mypy_path = "../jmd-impl"

[dependency-groups]
dev = [
    "mypy>=1.20.0",
    "pytest>=9.0.3",
    "ruff>=0.15.10",
]

